+++ src/utils/plugin-tools/index.ts
@@ -18,3 +18,6 @@ export { isJSON, toJSONString } from './config-tools';
 export { copyDataAttributes, createNode } from './dom-tools';
 
 export { sanitizeText } from './text-tools';
+
+export { addThemeColor } from './theme-tools';
+export type { ThemeColors, ThemeColorPair, ThemeColorOptions } from './theme-tools';